fixed a small error

+++ neuralNetworkGPU.h
@@ -32,7 +32,7 @@ private:
 public:
 
     //default constructor
-    NeuralNetworkGPU(float* valuesIn, int numNeuronInput, int numNeuronHidden, int numNeuronOutput,int numInputValuesX, int numInputValuesY, float learningRate);
+    NeuralNetworkGPU(int* valuesIn, int numNeuronInput, int numNeuronHidden, int numNeuronOutput,int numInputValuesX, int numInputValuesY, float learningRate);
 
     //train the neural network
     float* train(int numIterations, int tile_width);
